backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
python-multipart
python-jose[cryptography]
passlib[bcrypt]